[HTML5] Sticking plaster fix for https://bugs.opera.com/browse/GM-6783?filter=-1
The proper fix is to not modify the active lists whilst they are being iterated over. As per the C++ runner we should be building  a list of instances to activate\deactivate and only changing the lists when it is safe to do so. We should also implement a version of the WithObjIterator for consistency