Only use Spine collision mesh when selected (GM-6951).
Historically, we would apply the bounding box and/or collision mask
from the spine collision mesh or the precomputed mask from the sprite
editor inconsistently in different code paths.

In the April (beta) this got changed to always use the Spine collision
mesh, but then it was discovered people were relying on collisions
defined in the sprite editor even though they did not apply well to all
Spine sprites, so that change got backed out for the April stable.

This change adds a "Spine collision mesh" option to the sprite editor
which allows for explicitly choosing to use Spine collision meshes OR
GM collision masks in all cases on a per-sprite basis.