Allow use of spine collision meshes via mask_index (GM-7176).
If an instance's mask_index is set to a Spine sprite which is using a
collision mesh, use the mesh as it is in the sprite's default pose for
collisions.

This differs from our handling of per-frame precise collisions where we
select the mask from the mask_index sprite based on image_index, which
seems more sensible for Spine since there is a lot more state that can
differ between sprites (animations, attachments, etc).