Improve tracking of Spine animation wrapping.
The image_index instance variable is usually wrapped when it is
advanced automatically, and setAnimationTransform() attempted to
detect this through some heuristics, however with different
combinations of animation length, speed and position it could
incorrectly detect the animation wrapping with multiple outcomes,
such as the animation jumping backwards/forwards or running at the
wrong speed.

We now accumulate the "extra" frames when wrapping image_index and
add them to the index passed into setAnimationTransform() for a single
frame so setAnimationTransform() can accurately calculate the time
delta and advance the animation forwards/backwards correctly.

The heuristics are still in-place in case of any projects explicitly
setting the image_index on each frame but expecting the animation to
advance smoothly.

https://github.com/YoYoGames/GameMaker-Bugs/issues/1895