Add flag to skeleton_animation_clear() to reset skeleton state.
Removing an animation leaves any transforms/etc it applied to the
skeleton in place, this commit adds a "reset" argument which may be
passed to skeleton_animation_clear() to transform the skeleton back
to its setup pose. The duration parameter is optional, if not
specified the skeleton will snap back to its setup pose in one frame.

https://github.com/YoYoGames/GameMaker-Bugs/issues/1570