Add HTTP connect timeout function stubs.
We don't actually use the supplied connection timeout in the HTML5
runner, becase as far as I can tell, there is no way to set a timeout
specifically on the connection phase of an XmlHttpRequest.

The XmlHttpRequest 'timeout' member applies to the whole request, so it
will cause a timeout while still actively streaming data.

https://github.com/YoYoGames/GameMaker-Bugs/issues/5744