Fix for https://github.com/YoYoGames/GameMaker-Bugs/issues/12324
When a filter's temporary surface is drawn back to the main application surface it currently needs to cover the entire application surface, even if the effect only applies to a particular GameMaker viewport which covers only a section of the application surface. This is achieved by forcing the graphics viewport to cover the whole application surface before drawing the filter surface to it, but this was overriding any GameMaker viewport settings, which need to be restored afterwards (and what the code is now doing).